Rename '/public/' to '/media/' for Dodona 3.8.6
In Dodona 3.8.6 we've renamed the /public/ directory in the repository
root to /media/ to avoid confusion. This directory can be used to share
files between exercise descriptions. We've performed this rename commit
to fix repositories which would be broken by this release.

The /public/ directory now has a new purpose: files in this directory
are available to be used within course and series descriptions. There is
a listing of these files available on the repository page in Dodona.

For more info on the changed repository structure, you can visit the
updated documentation site:
https://dodona-edu.github.io/en/references/repository-directory-structure/